+++ spp_event_demo/wizard/create_event_phone_survey_wizard.py
@@ -8,6 +8,7 @@ class SPPCreateEventPhoneSurveyWizard(models.TransientModel):
     _description = "Create Event Phone Survey Wizard"
 
     event_id = fields.Many2one("spp.event.data")
+    target_type = fields.Selection([("individual", "Individual"), ("group", "Group")])
     summary = fields.Char()
     description = fields.Text()
